pyproject.toml
[build-system]
requires = ["setuptools"]

[project]
name = "notifications-utils"
dynamic = ["version"]
description = "Shared python code for GOV.UK Notify."
readme = "README.md"
license = "MIT"
authors = [
    { name = "Government Digital Service" },
]
dependencies = [
    "boto3[crt]>=1.34.100",
    "cachetools>=5.5.0",
    "Flask-Redis>=0.4.0",
    "Flask>=3.1.1",
    "govuk-bank-holidays>=0.17",
    "gunicorn[eventlet]>=23.0.0",
    "itsdangerous>=2.2.0",
    "Jinja2>=3.1.6",
    "mistune<2.0.0",  # v2 is totally incompatible with unclear benefit
    "ordered-set>=4.1.0",
    "phonenumbers>=9.0.9",
    "pypdf>=3.13.0",
    "python-dateutil>=2.9.0",
    "python-json-logger>=3.3.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.2",  # Can’t go past 2.32.2 until https://github.com/psf/requests/issues/6730 is fixed
    "segno>=1.6.1",
    "smartypants>=2.0.1",
    "statsd>=4.0.1",
]

[project.urls]
Homepage = "https://github.com/alphagov/notifications-utils"

[tool.pytest.ini_options]
xfail_strict = true
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.setuptools.dynamic]
version = {attr = "notifications_utils.version.__version__"}

[tool.setuptools.packages.find]
exclude = ["tests*"]
